+++ tests/testthat/test-rerun_lca.R
@@ -50,28 +50,30 @@ perform_test_lca <- function(settings, modeltypes){
   results
 }
 
-test_that('creates templates and performs rerun with new classes', {
-  get_path_from_type <- function(type){
-    paste0(settings$folder_name, '/', settings$analysis_name, '_model', type)
-  }
+if(is_mplus_installed()){
+  test_that('creates templates and performs rerun with new classes', {
+    get_path_from_type <- function(type){
+      paste0(settings$folder_name, '/', settings$analysis_name, '_model', type)
+    }
 
-  settings <- define_lca(random_testdata, 'test', 'id', nclasses = 2, starts = 5,
-                         use = c('var3', 'var4'))
-  results <- perform_test_lca(settings, modeltypes = c(1,2))
-  models_and_starts <- data.frame(modeltype = c(1,2),
-                                  classes = c(2,1),
-                                  starts = c(5,5))
-  capture.output(rerun_lca(results, models_and_starts))
-  for(i in seq(6)){
-    model_path <- paste0(get_path_from_type(i), '_template.txt')
-    expect_true(file.exists(model_path),
-                info = paste0('Did not write file for model ', i, '...'))
-  }
+    settings <- define_lca(random_testdata, 'test', 'id', nclasses = 2, starts = 5,
+                           use = c('var3', 'var4'))
+    results <- perform_test_lca(settings, modeltypes = c(1,2))
+    models_and_starts <- data.frame(modeltype = c(1,2),
+                                    classes = c(2,1),
+                                    starts = c(5,5))
+    capture.output(rerun_lca(results, models_and_starts))
+    for(i in seq(6)){
+      model_path <- paste0(get_path_from_type(i), '_template.txt')
+      expect_true(file.exists(model_path),
+                  info = paste0('Did not write file for model ', i, '...'))
+    }
 
-  out_path_12 <- paste0(get_path_from_type(1), '_lca/02_test_model1_lca.out')
-  out_path_21 <- paste0(get_path_from_type(2), '_lca/01_test_model2_lca.out')
-  expect_true(file.exists(out_path_12))
-  expect_true(file.exists(out_path_21))
-  unlink(settings$folder_name, recursive = T)
-})
+    out_path_12 <- paste0(get_path_from_type(1), '_lca/02_test_model1_lca.out')
+    out_path_21 <- paste0(get_path_from_type(2), '_lca/01_test_model2_lca.out')
+    expect_true(file.exists(out_path_12))
+    expect_true(file.exists(out_path_21))
+    unlink(settings$folder_name, recursive = T)
+  })
+}
